WoT S19a: extract cell/header DOM-access queries to table/cellAccess.ts
Move getCell/getRow/getColumnHeader(s)/getRowHeader(s)/getCoords/getTrForRow out
of table.ts into a cellAccess mixin applied once to base Table (inherited by all
subclasses, unlike the axis-selective range-query/sticky mixins). These methods
use no #-private state, so the move is a pure relocation onto `this`; behavior
unchanged. table.ts drops ~200 lines and the now-unused imports.

test:walkontable 745/0/2, test:types + eslint green.
